delet review

+++ app/reviews.py
@@ -1,3 +1,4 @@
+from fastapi import HTTPException, status
 from sqlalchemy.orm import Session
 from . import models, schemas
 
@@ -61,9 +62,21 @@ def create_review(db: Session,
 # Erase my review (by review Id)
 def delete_review(db: Session, user_id: int, rev_id: int):
     rev_db = db.query(models.Review).filter(
-    models.Review.id == rev_id).filter(
+    models.Review.id == rev_id,
     models.Review.userId == user_id
     ).first()
-    db.delete(rev_db)
-    db.commit()
-    return rev_db
+    if not rev_db:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Review not found."
+        )
+    try:
+       db.delete(rev_db)
+       db.commit()
+       return rev_db
+    except Exception as e:
+        db.rollback()
+        raise HTTPException(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            detail=f"An error occurred while deleting the review: {str(e)}"
+        )
